Impovements
Add Ability to add item, simply by dropping it into the Minimap Button.

Items in the GUI are now also showing the icon of item.

Add more functions to change default size and position of GUI. Not yet used.

Make Auto-Sell option default.

Remove Showing Item to Destroy as Minimap Button icon. (i think sometimes it's hard to find Button on minimap, as it was, because if you have like 5+ buttons its better to have static minimap icon)

To-Do:
- Make categories: Armor, Weapon, Misc, Trade Good, Consumable, etc... To show in GUI.
- Give user more configuration.